+++ Sources/iOS/TabViewController.swift
@@ -45,6 +45,9 @@ class TabViewController: TabmanViewController, PageboyViewControllerDataSource, 
         bar.buttons.customize {
             $0.tintColor = UIColor.tabmanForeground.withAlphaComponent(0.4)
             $0.selectedTintColor = .tabmanForeground
+            if #available(iOS 11, *) {
+                $0.adjustsFontForContentSizeCategory = true
+            }
         }
         bar.indicator.tintColor = .tabmanForeground